// src/providers/agritec/interface/agritec.cultivares.interface.ts
export interface AgritecCultivaresI {
  idCultivar: number;
  idCultura: number;
  safra: string;
  numeroRnc: string;
  obtentorMantenedor: string;
  cultivar: string;
  cultura: string;
  potencialProdutivo: number;
  duracaoCiclo: number;
  uf: string;
  grupo: string;
  maturacaoFisiologica: number;
  floracao: number;
  dataAtualizacao: string;
  regiao: string;
  grupoBioClimatico: string;
}