// src/providers/agritec/interface/agritec.produtividade.interface.ts
export interface AgritecResponseProdutividadeI {
  produtividadeAlmejada: number[];
  produtividadeMediaMunicipio: number[];
  temperaturaMinima: number[];
  temperaturaMaxima: number[];
  precipitacao: number[];
  grausDia: number[];
  balancoHidrico: number[];
  deficienciaHidrica: number[];
  excedenteHidrico: number[];
}
export interface AgritecProdutividadeI {
  produtividadeAlmejada: number;
  produtividadeMediaMunicipio: number;
  temperaturaMinima: number;
  temperaturaMaxima: number;
  precipitacao: number;
  grausDia: number;
  balancoHidrico: number;
  deficienciaHidrica: number;
  excedenteHidrico: number;
}